Move `SchemaCreation` to its own file instead of `AbstractAdapter`.
